pyproject.toml
[project]
name = "interviewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "llama-index-llms-anthropic>=0.3.4",
    "rich>=13.9.1",
    "tqdm>=4.66.5",
]
